medicine quantity - decimal field now

+++ ondoc/api/v1/prescription/serializers.py
@@ -49,7 +49,8 @@ class PrescriptionAppointmentValidation():
 class PrescriptionMedicineBodySerializer(serializers.Serializer):
     id = serializers.CharField(max_length=100)
     name = serializers.CharField(max_length=128)
-    quantity = serializers.IntegerField(required=False, allow_null=True)
+    # quantity = serializers.IntegerField(required=False, allow_null=True)
+    quantity = serializers.DecimalField(max_digits=10, decimal_places=2, required=False, allow_null=True)
     dosage_type = serializers.CharField(max_length=100, required=False, allow_blank=True)
     time = serializers.ListField(child=serializers.CharField(max_length=64), allow_empty=True, required=False)
     duration_type = serializers.ChoiceField(choices=prescription_models.PrescriptionMedicine.DURATION_TYPE_CHOICES, required=False, allow_null=True)
